turned phaser game into react component
