+++ package.json
@@ -25,7 +25,8 @@
     "colyseus.js": "^0.14.13",
     "express": "^4.18.2",
     "phaser": "^3.55.2",
-    "react": "^18.2.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "parcel": "^2.8.3",
