+++ src/client/components/App.tsx
@@ -1,0 +1,3 @@
+export function App() {
+  return <h1>Hello world!</h1>;
+}
