+++ src/client/index.html
@@ -7,7 +7,7 @@
     <title>Socket Shooter</title>
 </head>
 <body>
-
-    <script src="index.ts" type="module"></script>
+    <div id="root"></div>
+    <script src="index.tsx" type="module"></script>
 </body>
 </html>
