--- src/client/index.ts
@@ -1,18 +1,0 @@
-import Phaser from 'phaser';
-import { GameScene } from './scenes/GameScene';
-
-const config: Phaser.Types.Core.GameConfig = {
-  type: Phaser.AUTO,
-  width: 800,
-  height: 600,
-  physics: {
-    default: "arcade",
-    arcade: {
-      gravity: { y: 300 }
-    }
-  },
-  pixelArt: true,
-  scene: [ GameScene ],
-};
-
-const game = new Phaser.Game(config);
