+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
       "outDir": "./dist",
       "module": "commonjs",
-      "lib": ["es6"],
+      "lib": ["es6", "dom"],
       "target": "es2016",
       "declaration": true,
       "removeComments": true,
@@ -13,7 +13,8 @@
       "strict": false,
       "allowJs": true,
       "strictNullChecks": false,
-      "forceConsistentCasingInFileNames": true
+      "forceConsistentCasingInFileNames": true,
+      "jsx": "react"
   },
   "include": [
     "src"
